refactor: split SkiaRenderer class into different classes with different clear responsibilities
