+++ src/pictex/renderer/painters/__init__.py
@@ -1,0 +1,4 @@
+from .painter import Painter
+from .background import BackgroundPainter
+from .decoration import DecorationPainter
+from .text import TextPainter
