+++ src/pictex/renderer/painters/painter.py
@@ -1,0 +1,16 @@
+from abc import ABC, abstractmethod
+from ...models import Style
+from ..structs import RenderMetrics, Line
+from ..font_manager import FontManager
+import skia
+
+class Painter(ABC):
+    
+    def __init__(self, style: Style, metrics: RenderMetrics, font_manager: FontManager):
+        self._style: Style = style
+        self._metrics: RenderMetrics = metrics
+        self._font_manager: FontManager = font_manager
+
+    @abstractmethod
+    def paint(self, canvas: skia.Canvas, lines: list[Line]) -> None:
+        raise NotImplemented()
